+++ scripts/apps/archive/utils.ts
@@ -2,9 +2,7 @@
  * TODO: delete this file and update usages to use sdApi instead
  */
 
-import {sdApi} from 'api';
-import {appConfig} from 'appConfig';
-import {IArticle, IDesk, IStage} from 'superdesk-api';
+import {IArticle} from 'superdesk-api';
 import {PUBLISHED_STATES, KILLED_STATES, ITEM_STATE} from './constants';
 
 /**
@@ -22,21 +20,3 @@ export const isIngested = (item: IArticle) =>
  */
 export const isKilled = (item: IArticle) => KILLED_STATES.includes(item.state);
 
-export function getSendAndDuplicateTarget(): null | {deskId: IDesk['_id']; stageId: IStage['_id']} {
-    const desk = sdApi.desks.getAllDesks().find(
-        (desk) => desk.name === appConfig.features.customAuthoringTopbar.sendAndDuplicate.deskName,
-    );
-
-    if (desk == null) {
-        return null;
-    }
-
-    const stageFromConfig = sdApi.desks.getDeskStages(desk._id).find(
-        (desk) => desk.name === appConfig.features.customAuthoringTopbar.sendAndDuplicate.stageName,
-    );
-
-    return {
-        deskId: desk._id,
-        stageId: stageFromConfig != null ? stageFromConfig._id : desk.incoming_stage,
-    };
-}
